Adjusted the ArchitectConverter to use the drop system (and moved it to its own package).
